Combine csv-to-json parsers
Fixes #5
