+++ csv-to-json-all-years.ps1
@@ -1,0 +1,5 @@
+node ./2018/csv-to-json.js
+node ./2019/csv-to-json.js
+node ./2020/csv-to-json.js
+node ./2021/csv-to-json.js
+node ./2022/csv-to-json.js
